Fix toggle_checkbox mapping
For somereason, in orgmode, the org_mappings.toggle_checkbox does not
want to work with the normal mapping <C-Space>, so I had to create a
mapping that would directly bypass that mapping and instead would
just map <Leader><Space> directly to the lua function. That seemed to
work and I am going to go forward with that.

TODO
- Submit bug reprot to the orgmode.nvim github and see if there is
something that they can do.